SIG-1534 split up view code in V1 API
