+++ api/app/signals/apps/api/v1/views/__init__.py
@@ -1,0 +1,27 @@
+from signals.apps.api.v1.views.attachment import (
+    PrivateSignalAttachmentsViewSet,
+    PublicSignalAttachmentsViewSet
+)
+from signals.apps.api.v1.views.category import ChildCategoryViewSet, ParentCategoryViewSet
+from signals.apps.api.v1.views.category_removed import SignalCategoryRemovedAfterViewSet
+from signals.apps.api.v1.views.namespace import NamespaceView
+from signals.apps.api.v1.views.pdf import GeneratePdfView
+from signals.apps.api.v1.views.signal import PrivateSignalViewSet, PublicSignalViewSet
+from signals.apps.api.v1.views.signal_split import PrivateSignalSplitViewSet
+from signals.apps.api.v1.views.status_message_template import StatusMessageTemplatesViewSet
+from signals.apps.api.v1.views.stored_signal_filter import StoredSignalFilterViewSet
+
+__all__ = (
+    'PublicSignalAttachmentsViewSet',
+    'PrivateSignalAttachmentsViewSet',
+    'PublicSignalViewSet',
+    'PrivateSignalViewSet',
+    'SignalCategoryRemovedAfterViewSet',
+    'ChildCategoryViewSet',
+    'ParentCategoryViewSet',
+    'NamespaceView',
+    'GeneratePdfView',
+    'PrivateSignalSplitViewSet',
+    'StatusMessageTemplatesViewSet',
+    'StoredSignalFilterViewSet',
+)
